Refactors by moving error handlers to a module and adds Shorty class
