+++ src/app.rb
@@ -3,6 +3,10 @@ require 'sinatra'
 require 'json'
 require 'pry-byebug'
 require 'securerandom'
+require_relative 'modules/error_handler'
+require_relative 'modules/shorty'
+
+include ErrorHandler
 
 URL_PATTERN = /^[0-9a-zA-Z_]{4,}$/
 
@@ -33,10 +37,10 @@ post '/shorten' do
   return url_not_present if url.nil?
   return shortcode_no_match_pattern if !shortcode.nil? && !shortcode.match(URL_PATTERN)
   return shortcode_in_use unless settings.recorded_urls[shortcode].nil?
-  shortcode ||= generate_shortcode
-  settings.recorded_urls[shortcode] = {url: url, startDate: Time.now, lastSeenDate: nil, redirectCount: 0}
+  record = Shorty.new(url, shortcode, settings.recorded_urls).process
+  settings.recorded_urls[record[:shortcode]] = record
   status 201
-  {shortcode: shortcode}.to_json
+  {shortcode: record[:shortcode]}.to_json
 end
 
 get '/:shortcode/stats' do
@@ -46,30 +50,4 @@ get '/:shortcode/stats' do
   { startDate: record[:startDate],
     lastSeenDate: record[:startDate],
     redirectCount: record[:redirectCount] }.to_json
-end
-
-def url_not_present
-  status 400
-  {message: "'url' is not present"}.to_json
-end
-
-def shortcode_no_match_pattern
-  status 422
-  {message: "The shortcode fails to meet the following regexp: ^[0-9a-zA-Z_]{4,}$."}.to_json
-end
-
-def shortcode_in_use
-  status 409
-  {message: "The the desired shortcode is already in use. Shortcodes are case-sensitive."}.to_json
-end
-
-def shortcode_not_found
-  status 404
-  {message: "The shortcode cannot be found in the system"}.to_json
-end
-
-def generate_shortcode
-  shortcode = SecureRandom.hex(5)
-  return shortcode if settings.recorded_urls[shortcode].nil?
-  generate_shortcode
 end
